features profile picture by default not depending on user civility title done

+++ src/Controller/Admin/AdminUserController.php
@@ -52,6 +52,40 @@ class AdminUserController extends AbstractController
             // We call the uploadFile() method of the FileUploader service is order to upload the picture submit by the user.
             $picture = $this->fileUploader->uploadFile($form, 'picture');
 
+            //! START : profile picture by default without API.
+            // If we have a picture to upload.
+            if ($picture) {
+                // We set the picture to the user.
+                $user->setPicture($picture);
+            }
+            // Else the user not submit any picture so we set a picture by default.
+            else {
+                // We set the User::PICTURE to the user.
+                $user->setPicture(User::PICTURE);
+            }
+            //! END : profile picture by default without API.
+
+            //! START : profile picture by default according to user civility title without API.
+            // // If we have a picture to upload.
+            // if ($picture) {
+            //     // We set the picture to the user.
+            //     $user->setPicture($picture);
+            // }
+            // // Else the user not submit any picture so we set a picture by default depending on is gender.
+            // else {
+            //     // If the civility title of the user is User::MAN_CIVILITY_TITLE.
+            //     if ($user->getCivilityTitle() === User::MAN_CIVILITY_TITLE) {
+            //         // We set the User::MAN_PICTURE to the user.
+            //         $user->setPicture(User::MAN_PICTURE);
+            //     }
+            //     // Else if the civility title of the user is User::WOMAN_CIVILITY_TITLE.
+            //     elseif ($user->getCivilityTitle() === User::WOMAN_CIVILITY_TITLE) {
+            //         // We set the User::WOMAN_PICTURE to the user.
+            //         $user->setPicture(User::WOMAN_PICTURE);
+            //     }
+            // }
+            //! END : profile picture by default according to user civility title without API.
+
             //! START : profile picture by default with API.
             // If we have a picture to upload. 
             if ($picture) {
@@ -71,30 +105,7 @@ class AdminUserController extends AbstractController
 
                 // $user->setPicture($multiAvatarAPI->donwloadAvatar($multiAvatarAPI->fetch($user->getFirstName())));
             }
-
             //! END : profile picture by default with API.
-
-            //! START : profile picture by default without API.
-            // // If we have a picture to upload.
-            // if ($picture) {
-            //     // We set the picture to the user.
-            //     $user->setPicture($picture);
-            // }
-            // // Else the user not submit any picture so we set a picture by default depending on is gender.
-            // else {
-            //     // If the civility title of the user is User::MAN_CIVILITY_TITLE.
-            //     if ($user->getCivilityTitle() === User::MAN_CIVILITY_TITLE) {
-            //         // We set the User::MAN_PICTURE to the user.
-            //         $user->setPicture(User::MAN_PICTURE);
-            //     }
-            //     // Else if the civility title of the user is User::WOMAN_CIVILITY_TITLE.
-            //     elseif ($user->getCivilityTitle() === User::WOMAN_CIVILITY_TITLE) {
-            //         // We set the User::WOMAN_PICTURE to the user.
-            //         $user->setPicture(User::WOMAN_PICTURE);
-            //     }
-            // }
-            //! END : profile picture by default without API.
-
 
             // We call the persit() method of the EntityManagerInterface to put on hold the data.
             $this->entityManagerInterface->persist($user);
@@ -235,31 +246,45 @@ class AdminUserController extends AbstractController
             // We call the uploadFile() method of the FileUploader service is order to upload the picture submited by the user.
             $picture = $this->fileUploader->uploadFile($form, 'upload');
 
-            //! START : profile picture by default with API.
-            // If we have a picture to upload. 
-            if ($picture) {
-                // We get the current picture of the user that will be his previous picture after the switch. 
-                $previousPicture = $user->getPicture();
-
-                // We set to the picture property the value of $picture.
-                $user->setPicture($picture);
-
-                // We use the PHP function unlink() to delete, from our directory, the previous picture of the user. 
-                unlink($_ENV['USER_PICTURE_UPLOAD_FOLDER_PATH'] . '/' . $previousPicture);
-            }
-            // Else the user not submit any picture so we set a picture by default 
-            else {
-                // We don't want to change the picture that the user have already set for himself.
-                // We set the value of the initial user picture to the user.
-                $user->setPicture($user->getPicture());
-            }
-            //! END : profile picture by default with API.
-
             //! START : profile picture by default without API.
             // // If we have a picture to upload.
             // if ($picture) {
+            //     // We get the current picture of the user that will be his previous picture after the switch. 
+            //     $previousPicture = $user->getPicture();
+
             //     // We set the picture to the user.
             //     $user->setPicture($picture);
+
+            //     // If the previous picture of the user is different than User::PICTURE. 
+            //     if ($previousPicture !== User::PICTURE) {
+            //         // We use the PHP function unlink() to delete, from our directory, the previous picture of the user. 
+            //         unlink($_ENV['USER_PICTURE_UPLOAD_FOLDER_PATH'] . '/' . $previousPicture);
+            //     }
+            // }
+            // // Else the user not submit any picture so we set a picture by default.
+            // else {
+            //     // We set the value of the initial user picture to the user.
+            //     $user->setPicture($user->getPicture());
+            // }
+            //! END : profile picture by default without API.
+
+            //! START : profile picture by default according to user civility title without API.
+            // // If we have a picture to upload.
+            // if ($picture) {
+            //     // We get the current picture of the user that will be his previous picture after the switch. 
+            //     $previousPicture = $user->getPicture();
+
+            //     // We set the picture to the user.
+            //     $user->setPicture($picture);
+
+            //     // If the previous picture of the user is different than User::MAN_PICTURE and than User::WOMAN_PICTURE.
+            //     if (
+            //         $previousPicture !== User::MAN_PICTURE &&
+            //         $previousPicture !== User::WOMAN_PICTURE
+            //     ) {
+            //         // We use the PHP function unlink() to delete, from our directory, the previous picture of the user. 
+            //         unlink($_ENV['USER_PICTURE_UPLOAD_FOLDER_PATH'] . '/' . $previousPicture);
+            //     }
             // }
             // // Else the user not submit any picture so we set a picture by default depending on his gender.
             // else {
@@ -268,8 +293,10 @@ class AdminUserController extends AbstractController
             //         $user->getCivilityTitle() !== User::MAN_CIVILITY_TITLE || $user->getCivilityTitle() !== User::WOMAN_CIVILITY_TITLE
             //     ) {
             //         // If the user picture is different than User::MAN_PICTURE and than User::WOMAN_PICTURE.
-            //         if ($user->getPicture() !== User::MAN_PICTURE && 
-            //             $user->getPicture() !== User::WOMAN_PICTURE) {
+            //         if (
+            //             $user->getPicture() !== User::MAN_PICTURE &&
+            //             $user->getPicture() !== User::WOMAN_PICTURE
+            //         ) {
             //             // We don't want to change the picture that the user have already set for himself.
             //             // We set the value of the initial user picture to the user.
             //             $user->setPicture($user->getPicture());
@@ -289,8 +316,30 @@ class AdminUserController extends AbstractController
             //         }
             //     }
             // }
-            //! END : profile picture by default without API.
+            //! END : profile picture by default according to user civility title without API.
 
+            //! START : profile picture by default with API.
+            // If we have a picture to upload. 
+            if ($picture) {
+                // We get the current picture of the user that will be his previous picture after the switch. 
+                $previousPicture = $user->getPicture();
+
+                // We set to the picture property the value of $picture.
+                $user->setPicture($picture);
+
+                // If the previous picture of the user is different than User::PICTURE. 
+                if ($previousPicture !== User::PICTURE) {
+                    // We use the PHP function unlink() to delete, from our directory, the previous picture of the user. 
+                    unlink($_ENV['USER_PICTURE_UPLOAD_FOLDER_PATH'] . '/' . $previousPicture);
+                }
+            }
+            // Else the user not submit any picture so we set a picture by default 
+            else {
+                // We don't want to change the picture that the user have already set for himself.
+                // We set the value of the initial user picture to the user.
+                $user->setPicture($user->getPicture());
+            }
+            //! END : profile picture by default with API.
 
             // We call the flush() method of the EntityManagerInterface to backup the data in the database.
             $this->entityManagerInterface->flush();
@@ -335,6 +384,45 @@ class AdminUserController extends AbstractController
 
         // If the CSRF token is valid.
         if ($this->isCsrfTokenValid('delete-user-picture' . $user->getId(), $submittedToken)) {
+            //! START : profile picture by default without API.
+            // // / We get the current picture of the user that will be his previous picture after the switch. 
+            // $previousPicture = $user->getPicture();
+
+            // // We set the User::PICTURE to the user.
+            // $user->setPicture(User::PICTURE);
+
+            // // If the previous picture of the user is different than User::PICTURE. 
+            // if ($previousPicture !== User::PICTURE) {
+            //     // We use the PHP function unlink() to delete, from our directory, the previous picture of the user. 
+            //     unlink($_ENV['USER_PICTURE_UPLOAD_FOLDER_PATH'] . '/' . $previousPicture);
+            // }
+            //! END : profile picture by default without API.
+
+            //! START : profile picture by default according to user civility title without API.
+            // // We get the current picture of the user that will be his previous picture after the switch. 
+            // $previousPicture = $user->getPicture();
+
+            // // If the civility title of the user is User::MAN_CIVILITY_TITLE.
+            // if ($user->getCivilityTitle() === User::MAN_CIVILITY_TITLE) {
+            //     // We set the User::MAN_PICTURE to the user.
+            //     $user->setPicture(User::MAN_PICTURE);
+            // }
+            // // Else if the civility title of the user is User::WOMAN_CIVILITY_TITLE.
+            // elseif ($user->getCivilityTitle() === User::WOMAN_CIVILITY_TITLE) {
+            //     // We set the User::WOMAN_PICTURE to the user.
+            //     $user->setPicture(User::WOMAN_PICTURE);
+            // }
+
+            // // If the previous picture of the user is different than User::MAN_PICTURE and than User::WOMAN_PICTURE.
+            // if (
+            //     $previousPicture !== User::MAN_PICTURE &&
+            //     $previousPicture !== User::WOMAN_PICTURE
+            // ) {
+            //     // We use the PHP function unlink() to delete, from our directory, the previous picture of the user. 
+            //     unlink($_ENV['USER_PICTURE_UPLOAD_FOLDER_PATH'] . '/' . $previousPicture);
+            // }
+            //! END : profile picture by default according to user civility title without API.
+
             //! START : profile picture by default with API.
             // We get the current picture of the user that will be his previous picture after the switch. 
             $previousPicture = $user->getPicture();
@@ -347,26 +435,16 @@ class AdminUserController extends AbstractController
 
             // We set to the picture property the file name of the avatar. 
             $user->setPicture($avatar);
-            //! END : profile picture by default with API.
 
-            //! START : profile picture by default without API.
-            // // If the civility title of the user is User::MAN_CIVILITY_TITLE.
-            // if ($user->getCivilityTitle() === User::MAN_CIVILITY_TITLE) {
-            //     // We set the User::MAN_PICTURE to the user.
-            //     $user->setPicture(User::MAN_PICTURE);
-            // }
-            // // Else if the civility title of the user is User::WOMAN_CIVILITY_TITLE.
-            // elseif ($user->getCivilityTitle() === User::WOMAN_CIVILITY_TITLE) {
-            //     // We set the User::WOMAN_PICTURE to the user.
-            //     $user->setPicture(User::WOMAN_PICTURE);
-            // }
-            //! END : profile picture by default without API.
+            // If the previous picture of the user is different than User::PICTURE. 
+            if ($previousPicture !== User::PICTURE) {
+                // We use the PHP function unlink() to delete, from our directory, the previous picture of the user. 
+                unlink($_ENV['USER_PICTURE_UPLOAD_FOLDER_PATH'] . '/' . $previousPicture);
+            }
+            //! END : profile picture by default with API.
 
             // We call the flush() method of the EntityManagerInterface to backup the data in the database.
             $this->entityManagerInterface->flush();
-
-            // We use the PHP function unlink() to delete, from our directory, the previous picture of the user. 
-            unlink($_ENV['USER_PICTURE_UPLOAD_FOLDER_PATH'] . '/' . $previousPicture);
 
             // We display a flash message for the user.
             $this->addFlash('success', 'La photo de profil de ' . $user->getFirstName() . ' ' . strtoupper($user->getLastName()) . ' a bien été supprimée. Une photo de profil par défaut a été mise en place.');
